improved markdown support
