+++ init.rb
@@ -1,4 +1,6 @@
-require 'emojibutton_formatter_patch'
+require 'emojibutton_formatter'
+require 'emojibutton_formatter_textile_patch'
+require 'emojibutton_formatter_markdown_patch'
 require 'emojibutton_helper_patch'
 
 Redmine::Plugin.register :redmine_emojibutton do
@@ -14,9 +16,17 @@ Rails.configuration.to_prepare do
 
   # send Emoji Patches to all wiki formatters available to be able to switch formatter without app restart
   Redmine::WikiFormatting::format_names.each do |format|
-    unless Redmine::WikiFormatting::formatter_for(format).included_modules.include? EmojiButtonPlugin::Formatter::Patch
-      Redmine::WikiFormatting::formatter_for(format).send(:include, EmojiButtonPlugin::Formatter::Patch)
+    case format
+    when "markdown"
+      unless Redmine::WikiFormatting::Markdown::HTML.included_modules.include? EmojiButtonPlugin::Formatter::Markdown::Patch
+        Redmine::WikiFormatting::Markdown::HTML.send(:include, EmojiButtonPlugin::Formatter::Markdown::Patch)
+      end
+    when "textile"
+      unless Redmine::WikiFormatting::Textile::Formatter.included_modules.include? EmojiButtonPlugin::Formatter::Textile::Patch
+        Redmine::WikiFormatting::Textile::Formatter.send(:include, EmojiButtonPlugin::Formatter::Textile::Patch)
+      end
     end
+    
     unless Redmine::WikiFormatting::helper_for(format).included_modules.include? EmojiButtonPlugin::Helper::Patch
      Redmine::WikiFormatting::helper_for(format).send(:include, EmojiButtonPlugin::Helper::Patch)
     end
